requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.17.6
beautifulsoup4==4.11.1
lxml==4.8.0
ujson==5.3.0
httpx==0.22.0
requests==2.27.1
slowapi==0.1.5
python-dotenv==0.20.0